simplify behavioral data loaders
