--- hobj/data/behavior/oneshot_behavior.py
@@ -1,2 +1,0 @@
-#class MutatorOneShotDataset(LearningDataset):
-#    dataset_url = 'https://hlbdatasets.s3.us-east-1.amazonaws.com/behavior/mutator-oneshot-human-learning-data.json'
